Fix `setTimeout` memory leak of already unmounted components
The `attachListeners` functions already ensured that the SocketIO
ping interval gets cleared when using the `fromSocketIO` option,
but there was no guard for the reconnect timeouts defined in the
websocket event handler functions. This resulted in a memory leak when
a component that uses the `useWebsocket` React hook got unmounted right
after starting to connect to the defined websocket URL, but failed due
to any error like a invalid URL or error returned from the server.
The timed out functions defined in the `onerror` and/or `onclose`
handlers were still running resulting in network requests still being
fired and a "memory leak detection" error in the console thrown by
React.

This commit fixes this leak by clearing the timeouts in the React hook
cleanup function.

To reproduce the memory leak...

1. ...set the `reconnectAttempts` to any value larger than 0 or use the
   current default value of 5.
2. ...set the `reconnectInterval` to any value larger than 0 or use the
   current default value of 5000.
3. ...set the `retryOnError` to `true`.
4. ...pass a function to the `shouldReconnect` option that return
   `true`.
5. ...use the `useWebSocket` React hook in a component and mount it.
6. ...navigate to another view so that the previous component gets
   unmounted (e.g. using the `Link` component of the `@reach/router`
   library).
7. ...see the browser console and network request list (DevTools) to see
   the "memory leak detection" error being thrown as well as network
   request still being made towards the defined WebSocket URL.